refactor: split pybind11 bindings into domain-organized files
Split dal_bindings.cpp into bindings_core.cpp, bindings_global.cpp,
bindings_models.cpp, bindings_random.cpp, bindings_script.cpp,
bindings_value.cpp, and bindings_module.cpp, with a shared bindings.h
header. Each domain file exports an init_bindings_<domain> function
called by the PYBIND11_MODULE in bindings_module.cpp.
